Move content_types in grape folder since it's not a module Util
