+++ lib/grape.rb
@@ -218,7 +218,8 @@ module Grape
 end
 
 require 'grape/config'
-require 'grape/util/content_types'
+require 'grape/content_types'
+
 require 'grape/util/lazy_value'
 require 'grape/util/lazy_block'
 require 'grape/util/endpoint_configuration'
